Refactor of the code and some issues solved

+++ app/src/main/java/com/example/calculator/MathOperators.kt
@@ -1,21 +1,21 @@
 package com.example.calculator
 
 interface MathOperators {
-    fun doOperation(value1: Int, value2: Int): Int
+    fun doOperation(value1: Double, value2: Double): Double
 }
 
 class Sum(): MathOperators {
-    override fun doOperation(value1: Int, value2: Int) = value1 + value2
+    override fun doOperation(value1: Double, value2: Double) = value1 + value2
 }
 
 class Subtract(): MathOperators {
-    override fun doOperation(value1: Int, value2: Int) = value1 - value2
+    override fun doOperation(value1: Double, value2: Double) = value1 - value2
 }
 
 class Multiply(): MathOperators {
-    override fun doOperation(value1: Int, value2: Int) = value1 * value2
+    override fun doOperation(value1: Double, value2: Double) = value1 * value2
 }
 
 class Division(): MathOperators {
-    override fun doOperation(value1: Int, value2: Int) = value1 / value2
+    override fun doOperation(value1: Double, value2: Double) = value1 / value2
 }
